Update image description and README.md
The image description in the Github Actions workflow file has been updated to reflect that the Quarkus application is used for testing a CI/CD pipeline rather than being deployed using a Github Action. Notably, the change in README.md added information on using ArgoCD to deploy the application on a K3s Kubernetes cluster.